new practice

+++ Python 11/sport.py
@@ -1,4 +1,4 @@
-bet = int(input("Сделайте вашу ставку:"))
+bet = float(input("Сделайте вашу ставку:"))
 coef = float(input("Введите коэфицент:"))
 
 win = round(bet * coef,2)
